complete role permission

+++ database/migrations/2024_01_13_102338_create_commissions_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('commissions', function (Blueprint $table) {
             $table->id();
             $table->string('title');
-            $table->integer('commission');
+            $table->float('commission');
             
             $table->unsignedTinyInteger('status')->default(1)->comment('1= Active, 0= Inactive');
             $table->unsignedBigInteger('created_by')->nullable();
